[MISC] move type_traits/basic to utility
