+++ include/seqan3/alignment/configuration/align_config_scoring_scheme.hpp
@@ -16,9 +16,9 @@
 #include <seqan3/std/concepts>
 
 #include <seqan3/alignment/configuration/detail.hpp>
-#include <seqan3/core/type_traits/basic.hpp>
 #include <seqan3/alignment/scoring/scoring_scheme_concept.hpp>
 #include <seqan3/core/algorithm/pipeable_config_element.hpp>
+#include <seqan3/utility/type_traits/basic.hpp>
 
 namespace seqan3::align_cfg
 {
