+++ include/seqan3/alignment/matrix/detail/alignment_matrix_column_major_range_base.hpp
@@ -12,11 +12,12 @@
 
 #pragma once
 
-#include <seqan3/core/type_traits/basic.hpp>
-#include <seqan3/core/type_traits/range.hpp>
 #include <seqan3/std/iterator>
 #include <seqan3/std/ranges>
 #include <seqan3/std/span>
+
+#include <seqan3/core/type_traits/range.hpp>
+#include <seqan3/utility/type_traits/basic.hpp>
 
 namespace seqan3::detail
 {
