+++ include/seqan3/alignment/matrix/detail/alignment_trace_matrix_proxy.hpp
@@ -16,7 +16,7 @@
 
 #include <seqan3/alignment/matrix/trace_directions.hpp>
 #include <seqan3/core/simd/concept.hpp>
-#include <seqan3/core/type_traits/basic.hpp>
+#include <seqan3/utility/type_traits/basic.hpp>
 
 namespace seqan3::detail
 {
