+++ include/seqan3/alignment/matrix/detail/two_dimensional_matrix_iterator_base.hpp
@@ -12,13 +12,13 @@
 
 #pragma once
 
+#include <seqan3/std/concepts>
+#include <seqan3/std/ranges>
 #include <type_traits>
 
 #include <seqan3/alignment/matrix/detail/matrix_coordinate.hpp>
-#include <seqan3/core/type_traits/basic.hpp>
 #include <seqan3/core/type_traits/iterator.hpp>
-#include <seqan3/std/concepts>
-#include <seqan3/std/ranges>
+#include <seqan3/utility/type_traits/basic.hpp>
 
 namespace seqan3::detail
 {
