+++ include/seqan3/alignment/pairwise/policy/scoring_scheme_policy.hpp
@@ -14,7 +14,7 @@
 
 #include <seqan3/alphabet/concept.hpp>
 #include <seqan3/core/simd/concept.hpp>
-#include <seqan3/core/type_traits/basic.hpp>
+#include <seqan3/utility/type_traits/basic.hpp>
 
 namespace seqan3::detail
 {
