+++ include/seqan3/alphabet/detail/alphabet_proxy.hpp
@@ -12,13 +12,14 @@
 
 #pragma once
 
+#include <seqan3/std/concepts>
+
 #include <seqan3/alphabet/alphabet_base.hpp>
 #include <seqan3/alphabet/nucleotide/concept.hpp>
 #include <seqan3/alphabet/quality/concept.hpp>
 #include <seqan3/core/concept/core_language.hpp>
-#include <seqan3/core/type_traits/basic.hpp>
 #include <seqan3/core/type_traits/template_inspection.hpp>
-#include <seqan3/std/concepts>
+#include <seqan3/utility/type_traits/basic.hpp>
 
 namespace seqan3
 {
