+++ include/seqan3/core/algorithm/pipeable_config_element.hpp
@@ -17,7 +17,7 @@
 #include <type_traits>
 
 #include <seqan3/core/algorithm/concept.hpp>
-#include <seqan3/core/type_traits/basic.hpp>
+#include <seqan3/utility/type_traits/basic.hpp>
 
 namespace seqan3
 {
