+++ include/seqan3/core/concept/tuple.hpp
@@ -16,10 +16,10 @@
 #include <tuple>
 #include <type_traits>
 
-#include <seqan3/core/type_traits/basic.hpp>
-#include <seqan3/core/type_traits/template_inspection.hpp>
 #include <seqan3/core/pod_tuple.hpp>
+#include <seqan3/core/type_traits/template_inspection.hpp>
 #include <seqan3/utility/type_list/type_list.hpp>
+#include <seqan3/utility/type_traits/basic.hpp>
 
 namespace seqan3::detail
 {
