+++ include/seqan3/core/type_traits/all.hpp
@@ -18,9 +18,9 @@
 #pragma once
 
 #include <seqan3/core/type_traits/pre.hpp>
-#include <seqan3/core/type_traits/basic.hpp>
 #include <seqan3/core/type_traits/deferred_crtp_base.hpp>
 #include <seqan3/core/type_traits/iterator.hpp>
 #include <seqan3/core/type_traits/range.hpp>
 #include <seqan3/core/type_traits/template_inspection.hpp>
 #include <seqan3/core/type_traits/transformation_trait_or.hpp>
+#include <seqan3/utility/type_traits/basic.hpp>
