+++ include/seqan3/core/type_traits/range.hpp
@@ -12,14 +12,14 @@
 
 #pragma once
 
+#include <seqan3/std/iterator>
+#include <seqan3/std/ranges>
 #include <type_traits>
 
 #include <seqan3/core/platform.hpp>
 #include <seqan3/core/type_traits/pre.hpp>
-#include <seqan3/core/type_traits/basic.hpp>
 #include <seqan3/core/type_traits/iterator.hpp>
-#include <seqan3/std/ranges>
-#include <seqan3/std/iterator>
+#include <seqan3/utility/type_traits/basic.hpp>
 
 // TODO(h-2): add innermost_reference instead of or addition to range_innermost_value?
