+++ include/seqan3/range/views/char_to.hpp
@@ -12,10 +12,11 @@
 
 #pragma once
 
-#include <seqan3/alphabet/concept.hpp>
-#include <seqan3/core/type_traits/basic.hpp>
-#include <seqan3/range/views/deep.hpp>
 #include <seqan3/std/ranges>
+
+#include <seqan3/alphabet/concept.hpp>
+#include <seqan3/range/views/deep.hpp>
+#include <seqan3/utility/type_traits/basic.hpp>
 
 namespace seqan3::views
 {
