+++ include/seqan3/range/views/detail.hpp
@@ -13,11 +13,11 @@
 
 #pragma once
 
+#include <seqan3/std/ranges>
 #include <tuple>
 
-#include <seqan3/core/type_traits/basic.hpp>
 #include <seqan3/core/type_traits/range.hpp>
-#include <seqan3/std/ranges>
+#include <seqan3/utility/type_traits/basic.hpp>
 
 namespace seqan3::detail
 {
