+++ include/seqan3/range/views/get.hpp
@@ -12,9 +12,10 @@
 
 #pragma once
 
-#include <seqan3/core/type_traits/basic.hpp>
-#include <seqan3/core/concept/tuple.hpp>
 #include <seqan3/std/ranges>
+
+#include <seqan3/core/concept/tuple.hpp>
+#include <seqan3/utility/type_traits/basic.hpp>
 
 namespace seqan3::views
 {
