+++ include/seqan3/utility/type_traits/basic.hpp
@@ -14,7 +14,6 @@
 
 #include <tuple>
 #include <seqan3/std/type_traits>
-#include <type_traits>
 
 #include <seqan3/core/platform.hpp>
