+++ test/unit/core/my_tuple.hpp
@@ -15,7 +15,7 @@
 #include <tuple>
 #include <utility>
 
-#include <seqan3/core/type_traits/basic.hpp>
+#include <seqan3/utility/type_traits/basic.hpp>
 
 namespace seqan3
 {
