+++ test/unit/core/type_traits/basic_test.cpp
@@ -15,7 +15,7 @@
 #include <range/v3/view/iota.hpp>
 #include <range/v3/view/take_exactly.hpp>
 
-#include <seqan3/core/type_traits/basic.hpp>
+#include <seqan3/utility/type_traits/basic.hpp>
 
 //------------------------------------------------------------------------------
 // Tests for remove_cvref transformation trait
